+++ framework/core/__test__/mapper.test.ts
@@ -2,14 +2,12 @@
 import { SchemaValidator, number, string } from '@forklaunch/validator/typebox';
 import { PrimaryKey, Property } from '@mikro-orm/core';
 import { v4 } from 'uuid';
-import { RequestMapper } from '../src/mappers/models/requestMapper.model';
-import { ResponseMapper } from '../src/mappers/models/responseMapper.model';
+import { mapper } from '../src/mappers';
+import { BaseEntity } from '../src/persistence';
 
 const SV = SchemaValidator();
 
-type TypeboxSchemaValidator = typeof SV;
-
-class TestEntity {
+class TestEntity extends BaseEntity {
   @PrimaryKey({ type: 'uuid' })
   id: string = v4();
 
@@ -26,49 +24,31 @@ class TestEntity {
   age!: number;
 }
 
-class TestRequestMapper extends RequestMapper<
-  TestEntity,
-  TypeboxSchemaValidator
-> {
-  schema = {
+const TestMapper = mapper(
+  SV,
+  {
     id: string,
     name: string,
     age: number
-  };
-
-  async toEntity(arg1: string, arg2?: string): Promise<TestEntity> {
-    const entity = new TestEntity();
-    entity.id = this.dto.id;
-    entity.name = this.dto.name;
-    entity.age = this.dto.age;
-    return entity;
-  }
-}
-
-class TestResponseMapper extends ResponseMapper<
+  },
   TestEntity,
-  TypeboxSchemaValidator
-> {
-  schema = {
-    id: string,
-    name: string,
-    age: number
-  };
-
-  async fromEntity(
-    entity: TestEntity,
-    arg1: string,
-    arg2?: string
-  ): Promise<this> {
-    this.dto = {
-      id: entity.id,
-      name: entity.name,
-      age: entity.age
-    };
-
-    return this;
+  {
+    toEntity: async (dto, arg1: string, arg2?: string) => {
+      const entity = new TestEntity();
+      entity.id = dto.id;
+      entity.name = dto.name;
+      entity.age = dto.age;
+      return entity;
+    },
+    toDto: async (entity, arg1: string, arg2?: string) => {
+      return {
+        id: entity.id,
+        name: entity.name,
+        age: entity.age
+      };
+    }
   }
-}
+);
 
 function extractNonTimeBasedEntityFields<
   T extends {
@@ -86,14 +66,33 @@ function genericDtoWrapperFunction<T>(dto: T): T {
 }
 
 describe('request mappers tests', () => {
-  let TestRequestDM: TestRequestMapper;
+  let TestRequestDM: typeof TestMapper;
 
   beforeAll(() => {
-    TestRequestDM = new TestRequestMapper(SV);
+    TestRequestDM = {
+      schema: {
+        id: string,
+        name: string,
+        age: number
+      },
+      toEntity: async (dto, arg1: string, arg2?: string) => {
+        const entity = new TestEntity();
+        entity.id = dto.id;
+        entity.name = dto.name;
+        entity.age = dto.age;
+        return entity;
+      },
+      toDto: async (entity, arg1: string, arg2?: string) => {
+        return {
+          id: entity.id,
+          name: entity.name,
+          age: entity.age
+        };
+      }
+    };
   });
-
   test('schema static and constructed equality', async () => {
-    expect(TestRequestDM.schema).toEqual(TestRequestMapper.schema());
+    expect(TestRequestDM.schema).toEqual(TestMapper.schema);
   });
 
   test('from JSON', async () => {
@@ -103,17 +102,24 @@ describe('request mappers tests', () => {
       age: 1
     };
 
-    const responseDM = await TestRequestDM.fromDto(json);
-    const staticDM = await TestRequestMapper.fromDto(SV, json);
+    const responseDM = await TestRequestDM.toDto(
+      await TestRequestDM.toEntity(json, 'arg1'),
+      'arg1'
+    );
+    const staticDM = await TestMapper.toDto(
+      await TestMapper.toEntity(json, 'arg1'),
+      'arg1'
+    );
+
     const expectedDto = {
       id: '123',
       name: 'test',
       age: 1
     };
 
-    expect(staticDM.dto).toEqual(expectedDto);
-    expect(responseDM.dto).toEqual(expectedDto);
-    expect(responseDM.dto).toEqual(staticDM.dto);
+    expect(staticDM).toEqual(expectedDto);
+    expect(responseDM).toEqual(expectedDto);
+    expect(responseDM).toEqual(staticDM);
   });
 
   test('deserialize', async () => {
@@ -124,13 +130,13 @@ describe('request mappers tests', () => {
     };
 
     const entity = extractNonTimeBasedEntityFields(
-      await TestRequestDM.deserializeDtoToEntity(json, 'arg1')
+      await TestRequestDM.toEntity(json, 'arg1')
     );
-    const objectEntity = extractNonTimeBasedEntityFields(
-      await (await TestRequestDM.fromDto(json)).toEntity('arg1')
-    );
+    // const objectEntity = extractNonTimeBasedEntityFields(
+    //   await TestRequestDM.toEntity('arg1')
+    // );
     const staticEntity = extractNonTimeBasedEntityFields(
-      await TestRequestMapper.deserializeDtoToEntity(SV, json, 'arg1', 'arg2')
+      await TestMapper.toEntity(json, 'arg1', 'arg2')
     );
     let expectedEntity = new TestEntity();
     expectedEntity.id = '123';
@@ -140,12 +146,9 @@ describe('request mappers tests', () => {
     expectedEntity = extractNonTimeBasedEntityFields(expectedEntity);
 
     expect(entity).toEqual(expectedEntity);
-    expect(objectEntity).toEqual(expectedEntity);
     expect(staticEntity).toEqual(expectedEntity);
-    expect(entity).toEqual(objectEntity);
     expect(entity).toEqual(staticEntity);
     expect(staticEntity).toEqual(expectedEntity);
-    expect(staticEntity).toEqual(objectEntity);
   });
 
   test('serialize failure', async () => {
@@ -156,32 +159,44 @@ describe('request mappers tests', () => {
 
     await expect(
       async () =>
-        await TestRequestDM.fromDto(
+        await TestMapper.toEntity(
           // @ts-expect-error - missing age
-          json
-        )
-    ).rejects.toThrow();
-
-    await expect(
-      async () =>
-        await TestRequestMapper.fromDto(
-          SV,
-          // @ts-expect-error - missing age
-          json
+          json,
+          'arg1'
         )
     ).rejects.toThrow();
   });
 });
 
 describe('response mappers tests', () => {
-  let TestResponseDM: TestResponseMapper;
+  let TestResponseDM: typeof TestMapper;
 
   beforeAll(() => {
-    TestResponseDM = new TestResponseMapper(SV);
+    TestResponseDM = {
+      schema: {
+        id: string,
+        name: string,
+        age: number
+      },
+      toDto: async (entity, arg1: string, arg2?: string) => {
+        return {
+          id: entity.id,
+          name: entity.name,
+          age: entity.age
+        };
+      },
+      toEntity: async (dto, arg1: string, arg2?: string) => {
+        const entity = new TestEntity();
+        entity.id = dto.id;
+        entity.name = dto.name;
+        entity.age = dto.age;
+        return entity;
+      }
+    };
   });
 
   test('schema static and constructed equality', async () => {
-    expect(TestResponseDM.schema).toEqual(TestResponseMapper.schema());
+    expect(TestResponseDM.schema).toEqual(TestMapper.schema);
   });
 
   test('from entity', async () => {
@@ -190,17 +205,17 @@ describe('response mappers tests', () => {
     entity.name = 'test';
     entity.age = 1;
 
-    const responseDM = await TestResponseDM.fromEntity(entity, 'arg1');
-    const staticDM = await TestResponseMapper.fromEntity(SV, entity, 'arg1');
+    const responseDM = await TestResponseDM.toDto(entity, 'arg1');
+    const staticDM = await TestMapper.toDto(entity, 'arg1');
     const expectedDto = {
       id: '123',
       name: 'test',
       age: 1
     };
 
-    expect(staticDM.dto).toEqual(expectedDto);
-    expect(responseDM.dto).toEqual(expectedDto);
-    expect(responseDM.dto).toEqual(staticDM.dto);
+    expect(staticDM).toEqual(expectedDto);
+    expect(responseDM).toEqual(expectedDto);
+    expect(responseDM).toEqual(staticDM);
   });
 
   test('serialize', async () => {
@@ -210,13 +225,13 @@ describe('response mappers tests', () => {
     entity.age = 1;
 
     const json = genericDtoWrapperFunction(
-      await TestResponseDM.serializeEntityToDto(entity, 'arg1')
+      await TestResponseDM.toDto(entity, 'arg1')
     );
     const objectJson = genericDtoWrapperFunction(
-      await (await TestResponseDM.fromEntity(entity, 'arg1')).toDto()
+      await TestResponseDM.toDto(entity, 'arg1')
     );
     const staticJson = genericDtoWrapperFunction(
-      await TestResponseMapper.serializeEntityToDto(SV, entity, 'arg1', 'arg2')
+      await TestMapper.toDto(entity, 'arg1', 'arg2')
     );
     const expectedJson = {
       id: '123',
@@ -239,14 +254,7 @@ describe('response mappers tests', () => {
     entity.name = 'test';
 
     await expect(
-      async () =>
-        await (await TestResponseDM.fromEntity(entity, 'arg1')).toDto()
-    ).rejects.toThrow();
-    await expect(
-      async () =>
-        await (
-          await TestResponseMapper.fromEntity(SV, entity, 'arg1', 'arg2')
-        ).toDto()
+      async () => await TestMapper.toDto(entity, 'arg1', 'arg2')
     ).rejects.toThrow();
   });
 });