+++ framework/internal/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @forklaunch/common
 
+## 0.3.0
+
+### Minor Changes
+
+- remove class based mappers
+
+### Patch Changes
+
+- Updated dependencies
+  - @forklaunch/validator@0.10.0
+  - @forklaunch/common@0.6.0
+
+## 0.2.9
+
+### Patch Changes
+
+- add mappers as functions
+- Updated dependencies
+  - @forklaunch/validator@0.9.9
+  - @forklaunch/common@0.5.8
+
 ## 0.2.8
 
 ### Patch Changes